Extract NGram profile generation code to separate object, and put the implicit ordering in the package object.
