+++ src/main/scala/com/sixtysevenbricks/text/languagedetection/LanguageDetector.scala
@@ -21,16 +21,13 @@ import java.nio.file.Files
  */
 class LanguageDetector(fingerprintDir: File, languagesToCheck: List[String]) {
 
-  // We want higher counts to come first, not last; negating them is a cheap trick which works...
-  private implicit val tupleOrdering: Ordering[(String, Int)] = Ordering.by(x => (-x._2, x._1))
-
   private val fingerprints: Map[String, Seq[String]] = readFingerprints(fingerprintDir)
 
   /** Identify a text as being in a specific language, returning the name of that language. */
   def identifyLanguage(text: String): String = {
     val maxChars = 4
 
-    val profile = createNgramProfile(text, maxChars).take(300)
+    val profile = NGramProfiler.createNgramProfile(text, maxChars).take(300)
     val languageScores: Seq[(String, Int)] = for (language <- languagesToCheck;
                                                   languagePrints = fingerprints(language).filter(_.length <= maxChars))
     yield (language, computeDistance(profile, languagePrints))
@@ -46,27 +43,6 @@ class LanguageDetector(fingerprintDir: File, languagesToCheck: List[String]) {
     }.toMap
   }
 
-  /** Provide a total occurrence for each of the distinct terms in a sequence. i.e. (red,fish,blue,fish) will return: red 1, fish 2, blue 1. */
-  def countValues[T](terms: Seq[T]): Map[T, Int] = {
-    terms.groupBy(identity).myMapValues(_.length)
-  }
-
-  /** Extract the n-grams (bigrams, trigrams, etc.) from a piece of text into a sequence. Spaces are represented as _. */
-  def extractNgrams(text: String, maxChars: Int): Seq[String] = {
-    val normalizedText: String = "_" + text.toLowerCase.withNormalisedSpace.withoutPunctuation.replace(" ", "_") + "_"
-    for {start <- (0 to normalizedText.length);
-         length <- (1 to maxChars);
-         if start + length <= normalizedText.length} yield normalizedText.substring(start, start + length)
-  }
-
-  /** Provide a sorted summary count of the n-grams in a piece of text. */
-  def createNgramProfile(text: String, maxChars: Int): Profile =
-    sortNgramValues(countValues(extractNgrams(text, maxChars)))
-
-  /** Remove the unigrams, then sort the ngrams by their occurrence into a flat profile. */
-  private def sortNgramValues(countedValues: Map[String, Int]): Profile = {
-    countedValues.myFilterKeys(_.length > 1).toList.sorted.map(_._1)
-  }
 
   /** Work out the distance between two n-gram profiles. */
   private def computeDistance(ngrams: Profile, otherProfile: Profile): Int = {
